Fundamentally changed how the calculator operates:
Instead of displaying the expression in the lowerNum window,
lowerNum window now only displays the latest number entered,
which makes it easier for the posneg button to only affect
the last number entered.

When an operator is pressed, last number entered in lowerNum is
sent to upperNum plus the operator used. If last button clicked
was an operator, that operator is replaced with the latest one
to be clicked.

When equals is pressed, evaluates upperNum + lowerNum and displays
the result in lowerNum.

Consequently, updated the following :
- Clear button
- Number buttons
- Posneg button
- Operator buttons
- Equals button

Disabled the parentheses button until everything is in place to add
it back in.