Fixed number buttons logic - Reinstated max digits - Fixed behaviour when equalsClicked is true - Fixed issue with 0 button: when lowerNum is blank, can only enter 0 once, otherwise 0 will only be added if at least one other number is present.
Fixed equals button
- Added fixDoubleMinus function so that double minuses in upperNum
(ex. "6--7=") no longer create an error when equals is clicked
multiple times (secretly replaces "--" with "+" but doesn't change
display to show this).
- Added splitFirstNumber function to separate first number in upperNum
expression from rest of expression. This way, the rest of the expression
after the first number can be recombined with the result to update
upperNum.